Fixed endless overflow error
Due to network aberations, there are times with spikes of packets incoming. These cause internal buffers of encoded data to overrun and leak, which is normal and expected. Previously we would decode a null packet whenever a packet was lost, regardless of cause. This caused too much data to be decoded at one time, which in turn caused an endless overflow of buffers. Now we only decode a null packet when the network caused a packet to be lost